fx: Use proper integer types for brightness values
Backlight can go up to 96000 or more, use a u32. ddc values are 16-bit,
use u16.

+++ src/backlight.rs
@@ -2,7 +2,7 @@ use std::path::Path;
 
 use eyre::{Context, Result};
 
-pub fn backlight_brightness(path: &Path) -> Result<(u8, u8)> {
+pub fn backlight_brightness(path: &Path) -> Result<(u32, u32)> {
     let br_path = Path::new(path).join("brightness");
     let br =
         parse_path(br_path).with_context(|| format!("failed to read brightness for {:?}", path))?;
@@ -12,12 +12,12 @@ pub fn backlight_brightness(path: &Path) -> Result<(u8, u8)> {
     Ok((br, max_br))
 }
 
-pub fn set_backlight_brightness(path: &Path, new_br: u8) -> Result<(), eyre::Error> {
+pub fn set_backlight_brightness(path: &Path, new_br: u32) -> Result<(), eyre::Error> {
     let br_path = Path::new(path).join("brightness");
     std::fs::write(&br_path, new_br.to_string()).context("failed to write brightness")
 }
 
-fn parse_path(path: std::path::PathBuf) -> Result<u8> {
+fn parse_path(path: std::path::PathBuf) -> Result<u32> {
     std::fs::read_to_string(&path)?
         .trim()
         .parse()
